Add support for displaying donor messages and TTS messages
This commit adds support for displaying donor messages in the donation section of the webpage. If a donor message is provided, it will be shown along with the donor's name and the donated amount. Additionally, this commit also adds support for text-to-speech (TTS) messages. When a donation sound ends, the TTS message will be played after a 1-second delay, announcing the donor's name, donated amount, and donor message if provided.